Issue #17168: resources are disabled to compile by default, special cli paramter to is required to do it
