+++ .ci/validation.sh
@@ -509,6 +509,11 @@ check-since-version)
   fi
   ;;
 
+compile-test-resources)
+  # this task is useful during migration to new JDK to let compile resources on new jdk only
+  ./mvnw -e --no-transfer-progress clean test-compile -Dcheckstyle.skipCompileInputResources=false
+  ;;
+
 javac11)
   # InputCustomImportOrderNoPackage2 - nothing is required in front of first import
   # InputIllegalTypePackageClassName - bad import for testing
