+++ .circleci/config.yml
@@ -290,6 +290,10 @@ workflows:
           image-name: "cimg/openjdk:17.0.5"
           command: "./.ci/validation.sh javac17"
       - validate-with-script:
+          name: "java 17 test resources compile"
+          image-name: "cimg/openjdk:17.0.5"
+          command: "./.ci/validation.sh compile-test-resources"
+      - validate-with-script:
           name: "javac19"
           image-name: "cimg/openjdk:19.0.1"
           command: "./.ci/validation.sh javac19"
