+++ pom.xml
@@ -221,6 +221,7 @@
     </maven.sevntu-checkstyle-check.checkstyle.version>
     <maven.versions.plugin.version>2.17.1</maven.versions.plugin.version>
     <maven.compiler.plugin.version>3.14.0</maven.compiler.plugin.version>
+    <checkstyle.skipCompileInputResources>true</checkstyle.skipCompileInputResources>
     <java.version>11</java.version>
     <maven.compiler.release>${java.version}</maven.compiler.release>
     <pitest.plugin.version>1.19.4</pitest.plugin.version>
@@ -1476,6 +1477,7 @@
               <goal>add-test-source</goal>
             </goals>
             <configuration>
+              <skipAddTestSource>${checkstyle.skipCompileInputResources}</skipAddTestSource>
               <sources>
                 <source>src/test/resources</source>
               </sources>
@@ -1500,6 +1502,7 @@
               <goal>add-test-source</goal>
             </goals>
             <configuration>
+              <skipAddTestSource>${checkstyle.skipCompileInputResources}</skipAddTestSource>
               <sources>
                 <source>src/it/resources</source>
               </sources>
@@ -1524,6 +1527,7 @@
               <goal>add-test-source</goal>
             </goals>
             <configuration>
+              <skipAddTestSource>${checkstyle.skipCompileInputResources}</skipAddTestSource>
               <sources>
                 <source>src/xdocs-examples/resources</source>
               </sources>
