--- src/test/resources/com/puppycrawl/tools/checkstyle/packageobjectfactory/foo/FooCheck.java
@@ -1,20 +1,0 @@
-package com.puppycrawl.tools.checkstyle.packageobjectfactory.foo;
-
-import com.puppycrawl.tools.checkstyle.api.AbstractCheck;
-
-public class FooCheck extends AbstractCheck {
-    @Override
-    public int[] getDefaultTokens() {
-        return new int[] {0};
-    }
-
-    @Override
-    public int[] getAcceptableTokens() {
-        return getDefaultTokens();
-    }
-
-    @Override
-    public int[] getRequiredTokens() {
-        return getDefaultTokens();
-    }
-}
